pyproject.toml
[project]
run-etl = "run_pipeline.main:main"
name = "genaidrivenetl"
version = "0.1.0"
description = ""
authors = [
    {name = "Roman Devyatov",email = "romandevyatov98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "openai (>=2.16.0,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
